snapcove_backend/requirements.txt
# Django Core
Django==6.0

# Django REST Framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1

# Django Extensions
django-cors-headers==4.9.0
django-filter==24.3
django-celery-results==2.6.0

# Database
psycopg2-binary==2.9.11

# Image Processing
Pillow==12.0.0
exifread==3.0.0

# Machine Learning / AI
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0

# Task Queue
celery==5.3.4

# Authentication
google-auth==2.27.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0
